servers/fastapi-sqlalchemy/pyproject.toml
[tool.poetry]
name = "fastapi-sqlalchemy"
version = "0.1.0"
description = ""
authors = ["Doctor <thirvondukr@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = {extras = ["all"], version = "^0.70.0"}
SQLAlchemy = "^1.4.25"
alembic = "^1.7.4"
strawberry-graphql = "^0.83.0"
aiosqlite = "^0.17.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
